Fixed build error -  "authOptions" is not a valid Route export field.
