+++ src/app/signin/page.tsx
@@ -1,6 +1,6 @@
 import {SignInButton} from "@/components/auth/buttons";
 import {getServerSession} from "next-auth";
-import {authOptions} from "@/app/api/auth/[...nextauth]/route";
+import authOptions from "@/lib/auth/authOptions";
 import {getProviders} from "next-auth/react";
 import {redirect} from "next/navigation";
 import {Fragment} from "react";
